+++ udata_front/theme/gouvfr/assets/js/pages/OrganizationPublishingForm/Step1CreateOrJoinOrganization.vue
@@ -4,13 +4,13 @@
     <Container class="fr-mb-6v">
       <ActionCard
         :title="t('Join an Organization')"
-        :content="t('An organization is an entity in which many users can collaborate. The published datasets on the organization\'s name can be edited by its members')"
+        :content="t('An organization is an entity in which many users can collaborate. The published datasets on the organization\'s name can be edited by its members.')"
         :icon="organizationIcon"
         :stretchHeight="true"
       >
       <template #actions>
-        <div class="fr-col-12 fr-grid-row fr-grid-row--middle relative">
-          <div class="fr-col-12 fr-col-lg-9 fr-pb-2v">
+        <div class="fr-col-12 fr-grid-row fr-grid-row--top relative">
+          <div class="fr-col-12 fr-col-lg-9 fr-py-2v">
             <OrganizationSearch />
           </div>
           <div class="fr-col-12 fr-col-lg-3 fr-ml-auto">
@@ -35,7 +35,6 @@ import ActionCard from '../../components/Form/ActionCard/ActionCard.vue';
 import OrganizationSearch from '../../components/search/OrganizationSearch/OrganizationSearch.vue';
 import Stepper from '../../components/Form/Stepper/Stepper.vue';
 import Container from '../../components/Ui/Container/Container.vue';
-import { title } from "../../config";
 import organizationIcon from "../../../../templates/svg/illustrations/organization.svg";
 import { useI18n } from 'vue-i18n';